+++ src/components/Charts/PieChart.scss
@@ -1,3 +1,9 @@
+.chart{    
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
 .pie-chart{
     margin: 50px;
     display: flex;
