+++ src/components/Home/Home.jsx
@@ -1,31 +1,25 @@
 import React from 'react';
-import { useDispatch, useSelector } from 'react-redux'
-import { actionTest } from '../../redux/action'
+//import { useDispatch, useSelector } from 'react-redux'
+//import { actionTest } from '../../redux/action'
 import { useEffect } from 'react'
 import { useState } from 'react';
-import introImage from '../../assets/miscellaneous/home_wallpaper.png'
-import homeLogo from '../../assets/miscellaneous/home_logo.png'
 import './Home.scss'
 import PieChart from '../Charts/PieChart';
 import Aos from 'aos'
-import { Bot } from '../Chatbot/Chatbot'
-import 'aos/dist/aos.css'
-import config from '../Chatbot/Chatbot/config';
-import MessageParser from '../Chatbot/Chatbot/MessageParser';
-import ActionProvider from '../Chatbot/Chatbot/actionProvider';
 import TopPlayers from '../TopPlayers/TopPlayers';
 import { API_PORT } from '../../redux/constans'
 import ImageTest from './test.png'
-import Birds from './Birds/Birds';
-import Button from '../Chatbot/Button/Button';
+import Hero from '../Hero/Hero';
+import ChatbotContainer from '../Chatbot/ChatbotContainer';
+import 'aos/dist/aos.css'
 
-const Home = (props) => {
+const Home = () => {
 
     useEffect(() => {
         Aos.init({ duration: 1000, once: true })
+        // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [])
 
-    const [showBot, setShowBot] = useState(true)
     const [errorThree, setErrorThree] = useState(false)
     const [errorTen, setErrorTen] = useState(false)
     const [topthree, setTopthree] = useState([
@@ -116,44 +110,12 @@ const Home = (props) => {
         testingTopTen()
     }, [])
 
-    function showMeTheBot() {
-        const bot = document.getElementById('bot')
-        // eslint-disable-next-line 
-        { showBot ? setShowBot(false) : setShowBot(true) }
-        if (showBot) {
-            // bot.setAttribute('data-aos','fade-right');
-            bot.classList.add('showBot')
-            //   setAnimated(false)
-        }
-        else {
-            // bot.setAttribute('data-aos','');
-            bot.classList.remove('showBot')
-            //   setAnimated(true)
-        }
-    }
-
     return (
         <div className='home_container'>
-            
-            <Birds />
-
-            <div className="logo-location">
-                <img data-aos="fade" data-aos-duration={3000} src={homeLogo} alt="home_logo_img" id='home-logo' />
-            </div>
-            <img src={introImage} alt="home_img" id="home-wallpaper" />
-            <TopPlayers topthree={topthree} topten={topten}/>
-            <div className="chart">
-                <PieChart />
-            </div>
-
-            <div className="bot" id='bot' >
-                <Bot
-                    config={config}
-                    messageParser={MessageParser}
-                    actionProvider={ActionProvider}
-                />
-            </div>
-            <Button showMeTheBot={showMeTheBot} />
+            <Hero />
+            <TopPlayers topthree={topthree} topten={topten} />
+            <PieChart />
+            <ChatbotContainer />
         </div>
     );
 }
