+++ src/components/Home/Home.scss
@@ -4,36 +4,6 @@
     background: #0F5156;
 }
 
-#home-wallpaper{
-    width: 100%;
-    z-index: 1;
-    // position: absolute;
-}
-
-.logo-location{
-    display: flex;
-    justify-content: center;
-    width: 100%;    
-    align-items: center;
-}
-
-#home-logo{
-    width: 300px;
-    position: absolute;    
-    margin-top: 100px;
-    filter: drop-shadow(0px 10px 5px rgba(34, 34, 34, 0.39));
-}
-
-.chart{    
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-
-
-
-
 @import url('https://fonts.googleapis.com/css?family=Arima+Madurai:300');
 
 *,
